+++ Frameworks/OpenMPT/OpenMPT/contrib/fuzzing/get-afl.sh
@@ -1,10 +1,15 @@
 #!/usr/bin/env bash
 cd "${0%/*}"
-rm afl-latest.tgz
-wget http://lcamtuf.coredump.cx/afl/releases/afl-latest.tgz || exit
-tar -xzvf afl-latest.tgz
-rm afl-latest.tgz
-cd afl-*
+
+AFL_VERSION="$(wget --quiet -O - "https://api.github.com/repos/google/AFL/releases/latest" | grep -Po '"tag_name": "\K.*?(?=")')"
+AFL_FILENAME="$AFL_VERSION.tar.gz"
+AFL_URL="https://github.com/google/AFL/archive/$AFL_FILENAME"
+
+rm $AFL_FILENAME
+wget $AFL_URL || exit
+tar -xzvf $AFL_FILENAME
+rm $AFL_FILENAME
+cd AFL-*
 make || exit
 cd llvm_mode
 # may need to prepend LLVM_CONFIG=/usr/bin/llvm-config-3.8 or similar, depending on the system
@@ -13,4 +18,4 @@ cd ../libdislocator
 make || exit
 cd ../..
 rm -rf afl
-mv afl-* afl
+mv AFL-* afl