+++ Frameworks/OpenMPT/OpenMPT/soundlib/modcommand.cpp
@@ -308,6 +308,7 @@ void ModCommand::Convert(MODTYPE fromType, MODTYPE toType, const CSoundFile &snd
 			{
 				param = 0x0F | (std::min<PARAM>(0x0E, param & 0x0F) << 4);
 			}
+			break;
 
 		default:
 			break;